requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
PyMySQL==1.1.0
